Add support for nightfall widget in new structure
