+++ package.json
@@ -17,6 +17,7 @@
     "@stackla/jest-config": "^1.0.6",
     "@stackla/types": "^0.0.18-alpha.6",
     "@types/glidejs__glide": "^3.6.5",
+    "@types/masonry-layout": "^4.2.8",
     "@typescript-eslint/eslint-plugin": "^7.13.1",
     "esbuild": "^0.21.5",
     "eslint": "^8.56.0",
@@ -29,6 +30,7 @@
     "typescript": "^5.5.0"
   },
   "dependencies": {
-    "@glidejs/glide": "^3.6.1"
+    "@glidejs/glide": "^3.6.1",
+    "masonry-layout": "^4.2.2"
   }
 }
