+++ types/IBaseConfig.ts
@@ -1,0 +1,10 @@
+export interface BaseConfig {
+    tiles_per_page: number;
+    enable_custom_tiles_per_page: boolean;
+    rows_per_page: number;
+    click_through_url: string;
+    auto_refresh: boolean;
+    expanded_tile_show_shopspots: boolean;
+    expanded_tile_show_products: boolean;
+    expanded_tile_show_add_to_cart: boolean;
+}
