+++ widgets/libs/widget.components.ts
@@ -1,19 +1,13 @@
 import type { Sdk } from "@stackla/types";
-import { IWidgetSettings } from "types/IWidgetSettings";
-import { getConfig } from "./widget.config";
+import { BaseConfig } from "types/IBaseConfig";
 
 declare const sdk: Sdk;
 
-export function loadComponents() {
-  const widgetContainer = sdk.placement.getWidgetContainer();
-  const widgetSettings = getConfig(widgetContainer);
-
+export function loadExpandSettingComponents<T extends BaseConfig>(widgetSettings: T) {
   if (widgetSettings.expanded_tile_show_shopspots) {
     sdk.addLoadedComponents(["shopspots"]);
   }
-}
 
-export function loadExpandSettingComponents(widgetSettings: IWidgetSettings) {
   sdk.addLoadedComponents(["expanded-tile", "cross-sellers"]);
   if (widgetSettings.expanded_tile_show_products) {
     sdk.addLoadedComponents(["products"]);
