+++ widgets/nightfall/components/expanded-tile/base.css
@@ -1,0 +1,181 @@
+:host {
+  padding: 0;
+  box-sizing: border-box;
+  display: inline-block;
+  margin: 30px auto;
+  max-width: 1060px;
+  position: relative;
+  vertical-align: middle;
+  width: 100%;
+  display: flex;
+  flex-direction: row;
+  justify-content: center;
+}
+.exit {
+  position: absolute;
+  top: 20px;
+  right: 20px;
+  z-index: 1;
+  background: #fff;
+  border-radius: 50%;
+  width: 30px;
+  height: 30px;
+  display: flex;
+  flex-wrap: nowrap;
+  align-content: center;
+  justify-content: center;
+  align-items: center;
+}
+.panel {
+  position: relative;
+  background: #f4f4f4;
+  width: 600px;
+  height: 100%;
+  margin: 30px auto;
+  display: block;
+}
+.panel-left {
+  background: #fff;
+  display: flex;
+  align-content: center;
+  justify-content: center;
+  align-items: center;
+  width: 100%;
+  max-width: 100%;
+  min-width: 100%;
+
+}
+.panel-left img {
+  width: 100%;
+  display: block;
+}
+.panel-right {
+  padding: 24px 44px;
+  background: #2c2d2c;
+}
+.tile-arrows-left {
+  position: absolute;
+  width: 50px;
+  height: 50px;
+  top: 50%;
+  left: -60px;
+  border: 0;
+  background: none;
+}
+.tile-arrows-right {
+  position: absolute;
+  width: 50px;
+  height: 50px;
+  top: 50%;
+  right: -50px;
+  border: 0;
+  background: none;
+}
+
+.tile-arrows .widget-icon {
+  width: 50px;
+  height: 50px;
+  cursor: pointer;
+}
+div.image {
+  position: relative;
+}
+.image-wrapper {
+  position: relative;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  width: 100%;
+  margin-bottom: 0;
+}
+.image-wrapper-inner {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  width: 100%;
+}
+
+.image-wrapper-inner .image {
+  width: 100%;
+  display: block;
+}
+
+.user-info {
+  display: flex;
+  flex-direction: row;
+  gap: 10px;
+}
+
+.user-info-wrapper {
+  text-align: left !important;
+}
+
+.user-info-wrapper a {
+  color: #666;
+  font-family: 'Roboto Condensed', sans-serif;
+  text-decoration: none;
+  line-height: 1.4;
+}
+
+.user-top {
+  font-weight: 700;
+  font-size: 1.5em;
+  text-transform: uppercase;
+}
+
+.user-bottom {
+  font-size: 1.2em;
+  font-weight: 300;
+}
+
+.caption-paragraph {
+  color: #090909;
+  display: inline-block;
+  font-size: 12px;
+  line-height: 2;
+  text-align: left;
+  width: 100%;
+  word-break: break-word;
+}
+html {
+  background: --widget-background;
+}
+.ugc-tile {
+  width: --max-tile-width;
+  background: --text-tile-background;
+  margin-left: --margin;
+  margin-right: --margin;
+}
+.caption {
+  font-size: --text-tile-font-size;
+  color: --text-tile-font-color;
+}
+.content-inner-wrapper a {
+  color: --text-tile-link-color;
+}
+.user-name {
+  font-size: --text-tile-user-name-font-size;
+  color: --text-tile-user-name-font-color;
+}
+.user-handle {
+  font-size: --text-tile-user-handle-font-size;
+  color: --text-tile-user-handle-font-color;
+}
+.widget-icon {
+  display: block;
+}
+.tile-timestamp {
+  color: #ababab;
+  display: inline-block;
+  font-size: 12px;
+  padding-right: 20px;
+  text-align: left;
+  width: 100%;
+}
+.tile-caption {
+  margin-top: 15px;
+  display: --show_caption;
+  text-align: left;
+}
