+++ widgets/nightfall/components/products/base.css
@@ -1,0 +1,41 @@
+
+.stacklapopup-products-item-title-wrap {
+  text-decoration: none;
+}
+.stacklapopup-products-item-title {
+  color: #fff;
+  text-transform: uppercase;
+  font-size: 18px;
+  font-weight: bold;
+  line-height: 1.2;
+  margin-bottom: 10px;
+}
+.stacklapopup-products-item-price {
+  color: #fff;
+  font-size: 14px;
+  font-weight: normal;
+  margin-bottom: 5px;
+}
+.stacklapopup-product-images-wrapper {
+  max-width: 100%;
+}
+.stacklapopup-products-item-description {
+  color: #fff;
+  font-size: 14px;
+  font-weight: 400;
+}
+.stacklapopup-shopspot-cart,
+.stacklapopup-products-item-button {
+  color: --shopspot-btn_font_color;
+  background-color: --shopspot-btn_background;
+  border-radius: 4px;
+  display: inline-block;
+  font-size: 14px;
+  font-style: normal;
+  font-weight: bold;
+  padding: 15px 0px;
+  text-align: center;
+  text-decoration: none;
+  text-transform: uppercase;
+  min-width: 300px;
+}
