--- widgets/nightfall/components/shopspot-icon/base.styles.ts
@@ -1,36 +1,0 @@
-export default `
-  :host {
-      position: absolute;
-      top: 15px;
-      left: 15px;
-  }
-  
-  .stacklapopup-shopspot-cart-wrapper {
-      position: relative;
-  }
-  
-  .stacklapopup-shopspot-badge {
-      background: #fff;
-      border-radius: 50%;
-      cursor: pointer;
-      left: -12px;
-      position: absolute;
-      top: -12px;
-      padding: 5px 8px;
-      text-align: center;
-  }
-
-  .fs-tag {
-      background-image: url("data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' viewBox='0 0 512 512'%3E%3Cpath d='M345 39.1L472.8 168.4c52.4 53 52.4 138.2 0 191.2L360.8 472.9c-9.3 9.4-24.5 9.5-33.9 .2s-9.5-24.5-.2-33.9L438.6 325.9c33.9-34.3 33.9-89.4 0-123.7L310.9 72.9c-9.3-9.4-9.2-24.6 .2-33.9s24.6-9.2 33.9 .2zM0 229.5V80C0 53.5 21.5 32 48 32H197.5c17 0 33.3 6.7 45.3 18.7l168 168c25 25 25 65.5 0 90.5L277.3 442.7c-25 25-65.5 25-90.5 0l-168-168C6.7 262.7 0 246.5 0 229.5zM144 144a32 32 0 1 0 -64 0 32 32 0 1 0 64 0z' fill='%23fff' /%3E%3C/svg%3E");
-      background-repeat: no-repeat;
-      background-size: cover;
-      background-color: rgba(0,0,0,.6);
-      background-position: center;
-      height: 40px;
-      width: 40px;
-      color: white;
-      display: inline-block;
-      background-size: 70%;
-      border-radius: 50%;
-  }
-`;
