+++ widgets/nightfall/types/ISdkMasonry.ts
@@ -1,0 +1,6 @@
+import { Sdk } from "@stackla/types";
+import Masonry from "masonry-layout";
+
+export interface ISdkMasonry extends Sdk {
+    masonry: Masonry
+}
