+++ widgets/nightfall/widget.css
@@ -184,3 +184,9 @@
     background-size: contain;
     background-image: url("data:image/svg+xml,%3Csvg%20xmlns%3D%22http%3A%2F%2Fwww.w3.org%2F2000%2Fsvg%22%20width%3D%2224%22%20height%3D%2225%22%20viewBox%3D%220%200%2024%2025%22%20fill%3D%22none%22%3E%0A%3Crect%20width%3D%2224%22%20height%3D%2225%22%20rx%3D%223%22%20fill%3D%22black%22%2F%3E%0A%3Cpath%20d%3D%22M5.80574%2011.292C4.04586%209.29624%205.35838%206%207.91263%206H15.0931C16.6985%206%2018%207.40285%2018%209.13337V16.8604C18%2019.6155%2014.9383%2021.0295%2013.0878%2019.129L9.15153%2015.0868L5.80574%2011.292Z%22%20fill%3D%22%23FF66BF%22%2F%3E%0A%3C%2Fsvg%3E");
 }
+
+.tile-caption {
+    margin-top: 15px;
+    display: --show_caption;
+    text-align: left;
+  }
